Modified all locations where the employee model was referenced to fit the new user model
